requirements.txt
watchdog>=3.0.0
Pillow>=9.0.0
numpy>=1.20.0
opencv-python>=4.5.0
